requirements.txt
nemo_toolkit[asr]
torch
librosa
soundfile
wget
tqdm
numpy
transformers
pytorch-lightning